+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.31.2] - 2026-07-04
+
 ### Security
 
 - **Differenzierte Sichtbarkeit greift jetzt auch bei Statistik/Berichten.**